[AIRFLOW-5186] Move GCP Translate to core
This commit moves GCP Translate from contrib to core.
For more information check AIP-21.
