+++ docs/howto/operator/gcp/speech.rst
@@ -26,14 +26,14 @@ GcpTextToSpeechSynthesizeOperator
 Synthesizes text to audio file and stores it to Google Cloud Storage
 
 For parameter definition, take a look at
-:class:`airflow.contrib.operators.gcp_text_to_speech_operator.GcpTextToSpeechSynthesizeOperator`
+:class:`airflow.gcp.operators.text_to_speech.GcpTextToSpeechSynthesizeOperator`
 
 Arguments
 """""""""
 
 Some arguments in the example DAG are taken from the OS environment variables:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_speech.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_speech.py
       :language: python
       :start-after: [START howto_operator_text_to_speech_env_variables]
       :end-before: [END howto_operator_text_to_speech_env_variables]
@@ -43,14 +43,14 @@ google.cloud.texttospeech_v1.types module
 
 for more information, see: https://googleapis.github.io/google-cloud-python/latest/texttospeech/gapic/v1/api.html#google.cloud.texttospeech_v1.TextToSpeechClient.synthesize_speech
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_speech.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_speech.py
       :language: python
       :start-after: [START howto_operator_text_to_speech_api_arguments]
       :end-before: [END howto_operator_text_to_speech_api_arguments]
 
 filename is a simple string argument:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_speech.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_speech.py
       :language: python
       :start-after: [START howto_operator_text_to_speech_gcp_filename]
       :end-before: [END howto_operator_text_to_speech_gcp_filename]
@@ -58,7 +58,7 @@ filename is a simple string argument:
 Using the operator
 """"""""""""""""""
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_speech.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_speech.py
       :language: python
       :dedent: 4
       :start-after: [START howto_operator_text_to_speech_synthesize]
@@ -67,7 +67,7 @@ Using the operator
 Templating
 """"""""""
 
-.. literalinclude:: ../../../../airflow/contrib/operators/gcp_text_to_speech_operator.py
+.. literalinclude:: ../../../../airflow/gcp/operators/text_to_speech.py
     :language: python
     :dedent: 4
     :start-after: [START gcp_text_to_speech_synthesize_template_fields]
@@ -84,7 +84,7 @@ GcpSpeechToTextRecognizeSpeechOperator
 Recognizes speech in audio input and returns text.
 
 For parameter definition, take a look at
-:class:`airflow.contrib.operators.gcp_speech_to_text_operator.GcpSpeechToTextRecognizeSpeechOperator`
+:class:`airflow.gcp.operators.speech_to_text.GcpSpeechToTextRecognizeSpeechOperator`
 
 Arguments
 """""""""
@@ -94,14 +94,14 @@ google.cloud.speech_v1.types module
 
 for more information, see: https://googleapis.github.io/google-cloud-python/latest/speech/gapic/v1/api.html#google.cloud.speech_v1.SpeechClient.recognize
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_speech.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_speech.py
       :language: python
       :start-after: [START howto_operator_text_to_speech_api_arguments]
       :end-before: [END howto_operator_text_to_speech_api_arguments]
 
 filename is a simple string argument:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_speech.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_speech.py
       :language: python
       :start-after: [START howto_operator_speech_to_text_api_arguments]
       :end-before: [END howto_operator_speech_to_text_api_arguments]
@@ -109,7 +109,7 @@ filename is a simple string argument:
 Using the operator
 """"""""""""""""""
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_speech.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_speech.py
       :language: python
       :dedent: 4
       :start-after: [START howto_operator_speech_to_text_recognize]
@@ -118,7 +118,7 @@ Using the operator
 Templating
 """"""""""
 
-.. literalinclude:: ../../../../airflow/contrib/operators/gcp_speech_to_text_operator.py
+.. literalinclude:: ../../../../airflow/gcp/operators/speech_to_text.py
     :language: python
     :dedent: 4
     :start-after: [START gcp_speech_to_text_synthesize_template_fields]
