+++ docs/howto/operator/gcp/translate-speech.rst
@@ -30,7 +30,7 @@ GcpTranslateSpeechOperator
 Recognizes speech in audio input and translates it.
 
 For parameter definition, take a look at
-:class:`airflow.contrib.operators.gcp_translate_speech_operator.GcpTranslateSpeechOperator`
+:class:`airflow.gcp.operators.translate_speech.GcpTranslateSpeechOperator`
 
 Arguments
 """""""""
@@ -40,14 +40,14 @@ Config and audio arguments need to be dicts or objects of corresponding classes 
 
 for more information, see: https://googleapis.github.io/google-cloud-python/latest/speech/gapic/v1/api.html#google.cloud.speech_v1.SpeechClient.recognize
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_speech.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_speech.py
       :language: python
       :start-after: [START howto_operator_speech_to_text_api_arguments]
       :end-before: [END howto_operator_speech_to_text_api_arguments]
 
 Arguments for translation need to be specified.
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_speech.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_speech.py
       :language: python
       :start-after: [START howto_operator_translate_speech_arguments]
       :end-before: [END howto_operator_translate_speech_arguments]
@@ -56,7 +56,7 @@ Arguments for translation need to be specified.
 Using the operator
 """"""""""""""""""
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_speech.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_speech.py
       :language: python
       :dedent: 4
       :start-after: [START howto_operator_translate_speech]
@@ -65,7 +65,7 @@ Using the operator
 Templating
 """"""""""
 
-.. literalinclude:: ../../../../airflow/contrib/operators/gcp_translate_speech_operator.py
+.. literalinclude:: ../../../../airflow/gcp/operators/translate_speech.py
     :language: python
     :dedent: 4
     :start-after: [START translate_speech_template_fields]
