+++ tests/gcp/operators/test_speech_system_helper.py
